refactor: register_enclave_use_case! made smaller
