+++ example/key-vault/enclave/src/ecalls.rs
@@ -1,11 +1,7 @@
 use crate::ENCLAVE_CONTEXT;
-use anyhow::anyhow;
-use bincode::Options;
 use frame_enclave::{register_enclave_use_case, BasicEnclaveUseCase};
 use key_vault_ecall_types::cmd::*;
 use key_vault_enclave::{context::KeyVaultEnclaveContext, workflow::*};
-use log::error;
-use std::{ptr, vec::Vec};
 
 #[allow(dead_code)]
 struct DummyType;
