+++ frame/enclave/Cargo.toml
@@ -11,5 +11,8 @@ frame-runtime = { path = "../runtime", default-features = false, features = ["sg
 sgx_tstd = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git", features = ["net","backtrace", "untrusted_fs"] }
 anyhow = { rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/anyhow-sgx.git" }
 serde = { git = "https://github.com/mesalock-linux/serde-sgx.git", features = ["derive"] }
+bincode = { rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/bincode-sgx" }
 thiserror = { git = "https://github.com/mesalock-linux/thiserror-sgx.git" }
 sgx_types = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
+env_logger = { git = "https://github.com/mesalock-linux/env_logger-sgx", default-features = false, features = ["mesalock_sgx"] }
+log = { git = "https://github.com/mesalock-linux/log-sgx" }
