+++ frame/enclave/src/register_enclave_use_case.rs
@@ -8,6 +8,7 @@ macro_rules! register_enclave_use_case {
             ($cmd: path, $handler: ty),
         )*
     ) => {
+        #[cfg(feature = "runtime_enabled")]
         fn ecall_handler(cmd: u32, input: &mut [u8], ecall_max_size: usize) -> anyhow::Result<Vec<u8>> {
             match cmd {
                 $(
@@ -37,56 +38,22 @@ macro_rules! register_enclave_use_case {
             bincode::serialize(&res).map_err(Into::into)
         }
 
-        #[cfg(not(feature = "runtime_enabled"))]
-        fn inner_ecall_handler<EE>(input_payload: &[u8], ecall_max_size: usize) -> anyhow::Result<Vec<u8>>
-        where
-            EE: BasicEnclaveUseCase,
-        {
-            let res = {
-                let ecall_input = bincode::DefaultOptions::new()
-                    .with_limit(ecall_max_size as u64)
-                    .deserialize(&input_payload[..])
-                    .map_err(|e| anyhow!("{:?}", e))?;
-
-                EE::run::<$ctx_ops>(ecall_input, $ctx)?
-            };
-
-            bincode::serialize(&res).map_err(Into::into)
-        }
-
         #[no_mangle]
         pub extern "C" fn ecall_entry_point(
-            command: u32,
+            cmd: u32,
             input_buf: *mut u8,
             input_len: usize,
             output_buf: *mut u8,
             ecall_max_size: usize,
             output_len: &mut usize,
         ) -> frame_types::EnclaveStatus {
-            let input = unsafe { std::slice::from_raw_parts_mut(input_buf, input_len) };
-            let res = unsafe {
-                // TODO この箇所以外の前処理・後処理は、EnclaveUseCaseのデフォルト関数に任せる
-                match ecall_handler(command, input, ecall_max_size) {
-                    Ok(out) => out,
-                    Err(e) => {
-                        error!("Error in enclave (ecall_entry_point): command: {:?}, error: {:?}", command, e);
-                        return frame_types::EnclaveStatus::error();
-                    }
-                }
-            };
-
-            let res_len = res.len();
-            *output_len = res_len;
-
-            if res_len > ecall_max_size {
-                error!("Result buffer length is over ecall_max_size: ecall_max_size={}, res_len={}", ecall_max_size, res_len);
-                return frame_types::EnclaveStatus::error();
+            match cmd {
+                $(
+                    $(#[$feature])*
+                    $cmd => <$handler>::ecall_entry_point(input_buf,input_len,output_buf,ecall_max_size,output_len, $ctx),
+                )*
+                _ => unreachable!("Not registered the ecall command"),
             }
-            unsafe {
-                ptr::copy_nonoverlapping(res.as_ptr(), output_buf, res_len);
-            }
-
-            frame_types::EnclaveStatus::success()
         }
     }
 }
